feat: add shared email validator
