+++ apps/dentista/views/registro_dentista.py
@@ -1,30 +1,10 @@
 import flet as ft
-import re
 import requests
-from typing import Tuple
 from common import colors
+from common.validators import validar_email
 
 API_REGISTER_URL = "http://localhost:8000/api/usuarios/registrar"
 
-def validar_email(email: str) -> Tuple[bool, str]:
-    pattern = r'^[\w\.-]+@([\w\.-]+\.\w+)$'
-    match = re.match(pattern, email)
-    if not match:
-        return False, "Correo electrónico no válido"
-
-    dominio = email.split('@')[1].lower()
-    dominios_prohibidos = {
-        "gmai.com": "¿Quisiste decir gmail.com?",
-        "hotmial.com": "¿Quisiste decir hotmail.com?",
-        "yaho.com": "¿Quisiste decir yahoo.com?",
-        "outlok.com": "¿Quisiste decir outlook.com?",
-        "gmal.com": "¿Quisiste decir gmail.com?"
-    }
-
-    if dominio in dominios_prohibidos:
-        return False, dominios_prohibidos[dominio]
-
-    return True, ""
 
 def RegistroDentistaView(page: ft.Page):
     nombre_input = ft.TextField(label="Nombre", width=300, border_color=colors.PRIMARY, focused_border_color=colors.PRIMARY_DARK)
